# app/openapi.yaml
openapi: 3.1.0
info:
  title: CareerProxy API
  version: 1.0.0
  description: Job application consultancy platform API
servers:
  - url: https://careerproxy.com/api/v1
    description: Production
  - url: http://localhost:3000/api/v1
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Profile:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        full_name: { type: string }
        phone: { type: string }
        plan: { $ref: '#/components/schemas/SubscriptionPlan' }
        status: { type: string, enum: [active, paused, inactive] }
        role: { $ref: '#/components/schemas/UserRole' }
        onboarding_complete: { type: boolean }
        team_member_id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
...
    Application:
      type: object
      properties:
        id: { type: string, format: uuid }
        candidate_id: { type: string, format: uuid }
        team_member_id: { type: string, format: uuid }
        company: { type: string }
        job_title: { type: string }
        job_board: { $ref: '#/components/schemas/JobBoard' }
        job_url: { type: string }
        status: { $ref: '#/components/schemas/ApplicationStatus' }
        applied_at: { type: string, format: date-time }
        notes: { type: string }
    ApplicationStatus:
      type: string
      enum: [applied, viewed, interview_scheduled, rejected, offer, withdrawn]
    JobBoard:
      type: string
      enum: [LinkedIn, Indeed, Glassdoor]
    UserRole:
      type: string
      enum: [candidate, team_member, admin]
    SubscriptionPlan:
      type: string
      enum: [starter, pro, premium]
...
security:
  - BearerAuth: []
paths:
...
  /applications/bulk:
    post:
      summary: Bulk import applications (team only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                required: [candidate_id, company, job_title, job_board, job_url]
                properties:
                  candidate_id: { type: string }
                  company: { type: string }
                  job_title: { type: string }
                  job_board: { $ref: '#/components/schemas/JobBoard' }
                  job_url: { type: string }
                  applied_at: { type: string }
                  notes: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  imported: { type: integer }
                  failed: { type: integer }
                  errors: { type: array, items: { type: object } }
  /applications/{id}:
    get:
      summary: Get single application
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Application' }
    patch:
      summary: Update application status or notes
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { $ref: '#/components/schemas/ApplicationStatus' }
                notes: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Application' }
    delete:
      summary: Soft-delete application
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Deleted }
...
  /admin/candidates:
    get:
      summary: List all candidates
      parameters:
        - name: plan
          in: query
          schema: { $ref: '#/components/schemas/SubscriptionPlan' }
        - name: status
          in: query
          schema: { type: string }
        - name: team_member_id
          in: query
          schema: { type: string, format: uuid }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 25 }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Profile' }
                  total: { type: integer }
  /admin/team:
    get:
      summary: List all team members
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Profile' }
    post:
      summary: Create team member account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, full_name]
              properties:
                email: { type: string, format: email }
                full_name: { type: string }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
...